pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rw-core-keywords"
dynamic = ["version"]
description = "Core Robot Framework keyword libraries for the RunWhen platform"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [
    {name = "RunWhen", email = "engineering@runwhen.com"},
]

dependencies = [
    "PyGithub>=1.55",
    "PyYAML>=6.0",
    "prometheus-client>=0.11.0",
    "requests>=2.26.0",
    "robotframework>=4.1.2",
    "hvac>=0.11.2",
    "backoff>=2.2.1",
    "azure-identity>=1.14.0,<2.0.0",
    "azure-mgmt-containerservice>=23.0.0,<24.0.0",
    "azure-mgmt-resource>=23.0.0,<24.0.0",
    "azure-mgmt-subscription>=3.0.0,<4.0.0",
    "kubernetes>=30.1.0",
    "opentelemetry-api>=1.3.0",
    "opentelemetry-sdk>=1.3.0",
    "opentelemetry-exporter-otlp-proto-http>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.0",
]

[tool.setuptools.dynamic]
version = {file = "VERSION"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"RW" = ["AGENTS.md"]
